Spring security and adding authentication to end point

+++ Projects/JournelApp/src/main/java/com/JournelApp/JournelApp/services/JournalEntryService.java
@@ -26,7 +26,7 @@ public class JournalEntryService {
             User user = userService.findByUsername(username);
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }
         catch (Exception e) {
 //            System.out.println(e);
@@ -47,10 +47,20 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteJournalEntryById(ObjectId id, String username){
-        User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteJournalEntryById(ObjectId id, String username){
+        boolean removed =false;
+        try{
+            User user = userService.findByUsername(username);
+            removed =user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e){
+            throw new RuntimeException("Error deleting journal entry", e);
+        }
+        return removed;
+
     }
 }
